fixup! Move to DoFnTest
